Cell as separate component
