+++ src/App.tsx
@@ -1,43 +1,28 @@
-import { useState, useEffect, useRef, useCallback, ChangeEvent } from "react"
-import { Cell, CellRef } from "./types/types"
-import useKeyNavigation from "./hooks/useKeyNavigation"
-import useStoreCell from "./hooks/useStoreCell"
+import { useState, useCallback } from "react"
+import { CellRef } from "./types/types"
+import Cell from "./components/Cell"
 
 export const numRows = 25
 export const numCols = 10
 const rowSize = "30px"
 const colSize = "100px"
 
-const initialCells = () => {
-  const cells = []
-  for (let i = 0; i < numRows; i++) {
-    const row = []
-    for (let j = 0; j < numCols; j++) {
-      row.push({ value: "", styles: [] })
-    }
-    cells.push(row)
-  }
-  return cells
-}
+// const initialGrid = () => {
+//   const grid = []
+//   for (let i = 0; i < numRows; i++) {
+//     const row = []
+//     for (let j = 0; j < numCols; j++) {
+//       row.push({ row: i, col: j })
+//     }
+//     grid.push(row)
+//   }
+//   return grid
+// }
 
 function App() {
-  const [cells, setCells] = useState<Cell[][]>(initialCells())
+  // const [grid, setGrid] = useState<CellRef[][]>(initialGrid())
+  const grid: string[][] = Array(numRows).fill(Array(numCols).fill(""))
   const [activeCell, setActiveCell] = useState<CellRef>({ row: 0, col: 0 })
-  const [isEditing, setIsEditing] = useState<boolean>(false)
-
-  const inputRef = useRef<HTMLInputElement | null>(null)
-
-  const { setCellProperty, storeVal } = useStoreCell({ setCells, activeCell, setIsEditing })
-
-  useKeyNavigation({
-    activeCell,
-    setActiveCell,
-    cells,
-    setCellProperty,
-    isEditing,
-    setIsEditing,
-    storeVal,
-  })
 
   const colHeaders = useCallback(() => {
     const headers = []
@@ -53,29 +38,6 @@ function App() {
     return headers
   }, [])
 
-  useEffect(() => {
-    if (isEditing && inputRef.current) {
-      inputRef.current.focus()
-    }
-  }, [isEditing])
-
-  const isActiveCell = (row: number, col: number) => {
-    return activeCell.row === row && activeCell.col === col
-  }
-
-  const handleCellClick = (row: number, col: number) => {
-    if (isActiveCell(row, col)) {
-      setIsEditing(true)
-    } else {
-      storeVal()
-      setActiveCell({ row: row, col: col })
-    }
-  }
-
-  const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setCellProperty("value", e.target.value)
-  }
-
   return (
     <main>
       <table className="table-auto border-collapse">
@@ -88,38 +50,19 @@ function App() {
               </th>
             ))}
           </tr>
-          {cells.map((row, rowIndex) => (
+          {grid.map((row, rowIndex) => (
             <tr key={rowIndex}>
               <th className="border-y border-gray-500 px-1.5">{rowIndex + 1}</th>
-              {row.map((cell, colIndex) => (
-                <td key={colIndex} className="border  p-0 relative border-gray-500">
-                  <div
-                    className={`p-0.5 flex justify-center items-center relative ${cell.styles} ${
-                      activeCell.row === rowIndex && activeCell.col === colIndex
-                        ? "after:absolute after:-inset-px  after:content-[''] after:border-2 after:border-blue-500"
-                        : ""
-                    }`}
-                    style={{
-                      minWidth: colSize,
-                      minHeight: rowSize,
-                      width: colSize,
-                      height: rowSize,
-                    }}
-                    onClick={() => handleCellClick(rowIndex, colIndex)}
-                  >
-                    {isEditing && activeCell.row === rowIndex && activeCell.col === colIndex ? (
-                      <input
-                        value={cells[activeCell.row][activeCell.col].value}
-                        ref={inputRef}
-                        onChange={handleInputChange}
-                        onBlur={() => storeVal()}
-                        className="w-full h-full border-0 outline-none"
-                      />
-                    ) : (
-                      <p className={`w-full overflow-hidden`}>{cell.value}</p>
-                    )}
-                  </div>
-                </td>
+              {row.map((_, colIndex) => (
+                <Cell
+                  key={`${rowIndex}-${colIndex}`}
+                  rowIndex={rowIndex}
+                  colIndex={colIndex}
+                  rowSize={rowSize}
+                  colSize={colSize}
+                  activeCell={activeCell}
+                  setActiveCell={setActiveCell}
+                />
               ))}
             </tr>
           ))}
